Created a general layout of a webpage
This layout has a basic skeleton which I should add actual content and pictures to fill in.
Also, work on making the author of the quote move to the right of the quote on the bottom. Finally,
try to change how the header is laid out and use a ul and margin-right: auto to separate links and logo.